requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.1.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
albumentations=1.2.1=pypi_0
antlr4-python3-runtime=4.8=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
black=23.3.0=pypi_0
blinker=1.8.1=pypi_0
ca-certificates=2024.2.2=hbcca054_0
cachetools=5.3.2=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=2.1.0=pypi_0
comm=0.2.2=pyhd8ed1ab_0
configargparse=1.7=pypi_0
cycler=0.12.1=pypi_0
dash=2.16.1=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.8.1=py38h17151c0_0
decorator=5.1.1=pyhd8ed1ab_0
detectron2=0.6=pypi_0
docker-pycreds=0.4.0=pypi_0
executing=2.0.1=pyhd8ed1ab_0
fastjsonschema=2.19.1=pypi_0
filelock=3.13.1=pypi_0
fire=0.5.0=pypi_0
flask=3.0.3=pypi_0
fonttools=4.49.0=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.2.0=pypi_0
fvcore=0.1.5.post20220512=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.42=pypi_0
google-auth=2.28.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
grpcio=1.60.1=pypi_0
hydra-core=1.0.5=pypi_0
idna=3.6=pypi_0
imageio=2.34.0=pypi_0
importlib-metadata=3.10.1=pypi_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=7.1.0=hd8ed1ab_0
iopath=0.1.10=pypi_0
ipykernel=6.29.3=pyhd33586a_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.2=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pyhd8ed1ab_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonschema=4.21.1=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.7.2=py38h578d9bd_0
jupyterlab-widgets=3.0.10=pypi_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.5=pypi_0
krb5=1.21.2=h659d440_0
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libblas=3.9.0=21_linux64_openblas
libcblas=3.9.0=21_linux64_openblas
libedit=3.1.20191231=he28a2e2_2
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgfortran-ng=13.2.0=h69a702a_5
libgfortran5=13.2.0=ha4646dd_5
libgomp=13.2.0=h807b86a_5
liblapack=3.9.0=21_linux64_openblas
libopenblas=0.3.26=pthreads_h413a1c8_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=13.2.0=h7e041cc_5
lightning-utilities=0.11.2=pypi_0
loguru=0.6.0=pypi_0
markdown=3.3.4=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
minkowskiengine=0.5.4=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
natsort=8.3.1=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
networkx=3.1=pypi_0
ninja=1.11.1=pypi_0
numpy=1.24.4=py38h59b608b_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
omegaconf=2.0.6=pypi_0
open3d=0.18.0=pypi_0
opencv-python-headless=4.9.0.80=pypi_0
openssl=3.3.0=hd590300_0
packaging=23.2=pypi_0
pandas=2.0.3=py38h01efb38_1
parso=0.8.4=pyhd8ed1ab_0
pathspec=0.12.1=pypi_0
pathtools=0.1.2=pypi_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.2.0=pypi_0
pip=23.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.0=pypi_0
plotly=5.21.0=pypi_0
plyfile=0.7.4=pypi_0
pointnet2=0.0.0=pypi_0
portalocker=2.8.2=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.43=pypi_0
prompt_toolkit=3.0.42=hd8ed1ab_0
protobuf=3.19.6=pypi_0
psutil=5.9.8=py38h01eb140_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydeprecate=0.3.2=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pyquaternion=0.9.9=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-dotenv=0.20.0=pypi_0
python-tzdata=2024.1=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch-lightning=2.2.3=pypi_0
pytz=2024.1=pyhd8ed1ab_0
pyviz3d=0.2.28=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=26.0.3=py38ha44f8e3_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.0=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
retrying=1.3.4=pypi_0
rpds-py=0.18.0=pypi_0
rsa=4.9=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
sentry-sdk=1.40.5=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py38h06a4308_0
shortuuid=1.0.11=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tenacity=8.2.3=pypi_0
tensorboard=2.10.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.3.0=pypi_0
tifffile=2023.7.10=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torch=1.13.0+cu117=pypi_0
torchaudio=0.13.0+cu117=pypi_0
torchmetrics=0.11.4=pypi_0
torchvision=0.14.0+cu117=pypi_0
tornado=6.4=py38h01eb140_0
tqdm=4.66.2=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
trimesh=3.14.0=pypi_0
triton=2.2.0=pypi_0
typing-extensions=4.9.0=pypi_0
typing_extensions=4.11.0=pyha770c72_0
urllib3=2.2.1=pypi_0
volumentations=0.1.8=pypi_0
wandb=0.13.2=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.38.4=pypi_0
widgetsnbextension=4.0.10=pypi_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pypi_0
yarl=1.9.4=pypi_0
zeromq=4.3.5=h75354e8_4
zipp=3.17.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_0
